python_extractor/requirements.txt
fastapi
uvicorn
PyMuPDF
python-docx
pytesseract
pillow
python-multipart
